% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain-heart.R
\name{plotHEP}
\alias{plotHEP}
\title{Plot a heartbeat-evoked potential}
\usage{
plotHEP(x, main = "Heartbeat-evoked potential", ...)
}
\arguments{
\item{x}{A \code{"hep"} object.}

\item{main}{Title.}

\item{...}{Passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
\code{invisible(NULL)}.
}
\description{
Plot a heartbeat-evoked potential
}
