% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardioresp.R
\name{plotSynchrogram}
\alias{plotSynchrogram}
\title{Plot a cardiorespiratory synchrogram}
\usage{
plotSynchrogram(
  x,
  sampling_rate = NULL,
  main = "Cardiorespiratory synchrogram",
  ...
)
}
\arguments{
\item{x}{A \code{"synchrogram"} data frame or a \code{"cardioresp"} object.}

\item{sampling_rate}{Sampling rate, to label the x-axis in seconds (optional).}

\item{main}{Plot title.}

\item{...}{Passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
\code{invisible(NULL)}.
}
\description{
Plot a cardiorespiratory synchrogram
}
